Add support for `approle` auth method
This adds an approle authentication method in addition to the default
token-based auth method. It also adds configuration to select which
auth_method to use. The implementation of auth_method should make it
straight forward to add other autgh methods in the future.

Plus, this bumps the hvac min version to v0.10.6 (released 14 Dec 2020)
because the `client.auth.approle.login` method was added in that
release, deprecating the old `client.auth_approle` method.