drop unnecessary executable bit
The executable bit is a minor annoyance because it “lies” by saying all
files are executable when yaml and images are not. If everything is
executable then the `ls` coloring is also screwed up which slows me down
because I wonder why everything is showing as executable instead of as
plain text, etc.

Eventually I also want to fix the pack install so it stops marking
everything executable.