Work around TLS errors on CircleCI
For some reason the ssl module is hitting an infinite recursion error.
This might be CircleCI-specific, depending on the openssl lib they used
to compile python. The 3.7.0 version is supposed to be even more robust,
so it could also be something about v3.6.13. I have not reproduced the
error elsewhere, however.

The error began appearing after we updated requests from v2.23.0 to
v2.25.1 in https://github.com/StackStorm/st2/pull/5215
In requests v2.24.0, it defaulted to using the python ssl module instead
of PyOpenSSL due to various SELinux issues. Before that it used
PyOpenSSL whenever it was installed. Since it worked with pyopenssl, we
just revert to using it for the tests in this repo.
See: https://github.com/psf/requests/pull/5443